hola

+++ P4/pescalar_serie.c
@@ -14,7 +14,7 @@ int main(int argc, char **argv)
 	float sum=0;
 	unsigned long long tam;
 
-	tam = atoi(argv[1]);
+	tam = atoll(argv[1]);
 	
 	A = generateVector(tam);
 	B = generateVector(tam);
